// templates/salesforce/types.ts
interface Attributes {
    type: string;
    url: string;
}
// ...
export interface SalesforceContact {
    attributes: Attributes;
    Id: string;
    FirstName: string | null;
    MiddleName: string | null;
    LastName: string;
    Account: NamedEntity | null;
    Email: string | null;
    AccountId: string | null;
    OwnerId: string;
    Owner: NamedEntity;
    MobilePhone: string | null;
    Phone: string | null;
    Title: string | null;
    Salutation: string | null;
    LastModifiedDate: string;
}

export interface SalesforceArticle {
    Id: string;
    Title: string;
    LastModifiedDate: string;
    [key: string]: any;
}

export interface SalesforceDeal {
    attributes: Attributes;
    Id: string;
    Name: string;
    Amount: number | null;
    StageName: string;
    AccountId: string;
    LastModifiedDate: string;
}

interface Attributes {
    type: string;
    url: string;
}

interface NamedEntity {
    attributes: Attributes;
    Name: string;
}